requirements.txt
python-telegram-bot
openai
pandas
psycopg2-binary
matplotlib
